+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 # Changelog
 * Added support for the incident fields in outgoing mappers in the ID set.
 * Added a validation that the taskid field and the id field under the task field are both from uuid format and contain the same value.
+* Added a validation that an integration command has at most one default argument.
 
 # 1.2.17
 * Added a validation that the classifier of an integration exists.